+++ src/main/java/com/vedatech/pro/model/supplier/Supplier.java
@@ -7,6 +7,7 @@ import lombok.Getter;
 import lombok.Setter;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -18,8 +19,8 @@ public class Supplier extends ContactInfo
     private String company;
     private String displayName;
     private Boolean status;
-    private Double balance;
-    private Double balanceToday;
+    private BigDecimal balance;
+    private BigDecimal balanceToday;
     @Column(name = "supplier_rfc")
     private String supplierRfc;
