+++ src/main/java/com/vedatech/pro/repository/bank/BankTransactionDao.java
@@ -5,14 +5,15 @@ import com.vedatech.pro.model.bank.BankTransaction;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface BankTransactionDao extends CrudRepository<BankTransaction, Long> {
 
 
     @Query("SELECT SUM(e.deposito) - SUM(e.retiro) FROM BankTransaction e where e.bank.id = ?1")
-    Double getTotalBankTransBalance(Long id);
+    BigDecimal getTotalBankTransBalance(Long id);
 
     @Query("SELECT SUM(e.deposito) - SUM(e.retiro) FROM BankTransaction e where e.bank.subAccount = ?1")
-    Double getBalanceBySubAccount(Long id);
+    BigDecimal getBalanceBySubAccount(Long id);
 }
